fix(nn): make EMA's decay update round identically in all three shapes
TestEMAUpdateBitIdenticalToSlowPath fails on arm64 by one ulp. Unlike the four
fused-path fixes before it, this one is NOT fused-versus-dispatch: it is the
same formula written three ways, and arm64 contracts each differently.

  fast f64 arm   a slice loop            avg[i] = Decay*avg[i] + (1-Decay)*d[i]
  fast f32 arm   the same, plus a conversion on the operand
  generic path   a CLOSURE passed to readGen
  test oracle    a plain loop indexing e.avg[pi][i] two levels deep

`x*a + y*b` has two multiplies and arm64 fuses ONE of them into the add. Which
one it picks depends on the surrounding expression, so a slice loop, a closure
body and a two-level index do not agree — on amd64 they all agree, because
nothing contracts. Rounding both products explicitly makes all four shapes
compute the same number on every architecture.

THIS EDITS THE TEST'S REFERENCE, which the four previous fixes did not, and that
is deliberate rather than convenient: the oracle is a restatement of the same
formula, so it has to state the same rounding for the comparison to mean
anything. Left bare it pins the implementation to whatever arm64 happens to fuse
in a test helper.

nn failures 6 -> 5.

TWO ATTEMPTS THAT DID NOT WORK, recorded so they are not retried. Forcing
no-contraction in the implementation alone moved the failing element without
fixing it — the oracle still contracted. Hoisting the f32 arm's conversion into
a variable so its shape matched the generic path did not fix it either; the
generic path is a closure, and matching the source text is not the same as
matching what the compiler emits.

THE FIVE THAT REMAIN are fused-versus-DISPATCH, where only the Go side is
editable and the backend kernels contract too, so suppressing contraction on one
side alone moves the divergence rather than removing it — verified on GLA, where
it shifted the first mismatch from element 49 to element 1 and was reverted
rather than shipped. Those need a decision about which side is normative.

+++ nn/weightavg.go
@@ -123,20 +123,20 @@ func (e *EMA) Update() error {
 			case tensor.F64:
 				d := p.Storage().F64()
 				for i := range avg {
-					avg[i] = e.Decay*avg[i] + (1-e.Decay)*d[i]
+					avg[i] = float64(e.Decay*avg[i]) + float64((1-e.Decay)*d[i])
 				}
 				continue
 			case tensor.F32:
 				d := p.Storage().F32()
 				for i := range avg {
-					avg[i] = e.Decay*avg[i] + (1-e.Decay)*float64(d[i])
+					avg[i] = float64(e.Decay*avg[i]) + float64((1-e.Decay)*float64(d[i]))
 				}
 				continue
 			}
 		}
 		idx := 0
 		readGen(p, func(w float64) {
-			avg[idx] = e.Decay*avg[idx] + (1-e.Decay)*w
+			avg[idx] = float64(e.Decay*avg[idx]) + float64((1-e.Decay)*w)
 			idx++
 		})
 	}
